\name{clusterindices-package}
\alias{clusterindices-package}
\docType{package}
\title{
Cluster Validity Indices
}

\description{
Numerous indices to choose the optimal number of clusters when performing k-means.
}

\details{
\tabular{ll}{
Package: \tab clusterindices\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2026-06-05\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr} and Nikolaos Kontemeniotis \email{kontemeniotisn@gmail.com}.
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>
}

\references{
Tsagris M. and Kontemeniotis N. (2025). Simplicial clustering using the \eqn{\alpha}--transformation.
Lobachevskii Journal of Mathematics, 46: 6471–-6482.
https://arxiv.org/pdf/2509.05945.


Garcia-Escudero Luis A., Gordaliza Alfonso, Matran Carlos, Mayo-Iscar Agustin. (2008). A general
trimming approach to robust cluster analysis. Annals of Statistics 36(3): 1324--1345.
}

%\note{
%}


